feat: add show_broken_pkg reference implementation
This commit also adds the following:

dep_state method for the depcache to get the state of a dependency
DepFlags from apt to bitwise with the dep_state
is_critical on the dependency to determine if it's considered a critical dependency type.
is_important_dep on the depcache to determine if a dep is considered important. All critical deps are also important deps.